Mplus Print Input Instructions

Show the input instructions.

Usage

mplus.print.input.instructions(file, quiet = FALSE)

Arguments

file

the quoted name of an existing H5 file

quiet

logical

Value

The Mplus input instructions are printed to the screen

Examples

ex3_1 <- system.file("extdata", "ex3_1.h5", package = "mplush5")
mplus.print.input.instructions(ex3_1)
#> TITLE: 
#> this is an example of a simple linear 
#> regression for a continuous observed 
#> dependent variable with two covariates 
#> DATA: 
#> FILE = "ex3_1_aa78f01013fe7acb644cf7c3710aa050.dat"; 
#>  
#> VARIABLE: 
#> NAMES = y1 x1 x3; 
#> MISSING=.; 
#>  
#> MODEL: 
#> y1 ON x1 x3; 
#> OUTPUT: 
#> standardized; cinterval; 
#> SAVEDATA: 
#> H5RESULTS = ex3_1.h5; 

ex5_5 <- system.file("extdata", "ex5_5.h5", package = "mplush5")
mplus.print.input.instructions(ex5_5)
#> TITLE: 
#> this is an example of a two-parameter logistic 
#> item response theory (IRT) model 
#> DATA: 
#> FILE = "ex5_5_0e06abc9e90586008564a284d1c653ea.dat"; 
#>  
#> VARIABLE: 
#> NAMES = u1 u2 u3 u4 u5 u6 u7 u8 u9 u10 u11 u12 u13 u14 u15 u16 u17 u18 u19 u20; 
#> MISSING=.; 
#> CATEGORICAL ARE u1-u20; 
#> ANALYSIS: 
#> ESTIMATOR = MLR; 
#> MODEL: 
#> f BY u1-u20*; 
#> f@1; 
#> OUTPUT: 
#> standardized; cinterval; TECH1; TECH8; 
#> SAVEDATA: 
#> H5RESULTS = ex5_5.h5;