mplush5: Extract the Results Stored in the H5 File Saved From Mplus

This package contains a set of functions that extract the results stored in the H5 file saved by Mplus. The functions were written by Thuy Nguyen of Muthen & Muthen and can be found at https://www.statmodel.com/mplus-R/H5Results.zip. Minimal modifications were made to make them work in a package.

See also

Author

Maintainer: Douglas Tommet douglas_tommet@brown.edu